// src/services/wallet-state-core/generated/NetView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NetEndpointView } from "./NetEndpointView";
import type { NetNetworkRow } from "./NetNetworkRow";
import type { NetProviderView } from "./NetProviderView";
import type { NetWizardView } from "./NetWizardView";

export type NetView = { loaded: boolean, networks: Array<NetNetworkRow>, wizard: NetWizardView, endpoints: Array<NetEndpointView>, providers: Array<NetProviderView>, last_added_chain_id: number | null, };